refactor: merge different validate funtions
merge validateLogConfig into validateConfig, and move validate functions
into file daemon/mgr/container_validation.go
